+++ R/mcmc-diagnostics.R
@@ -170,8 +170,8 @@ mcmc_rhat <- function(rhat, ..., size = NULL) {
     labs(y = NULL, x = expression(hat(R))) +
     scale_fill_diagnostic("rhat") +
     scale_color_diagnostic("rhat") +
-    scale_x_continuous(breaks = brks, expand = c(0, .01)) +
-    scale_y_discrete(expand = c(.025,0)) +
+    scale_x_continuous(breaks = brks, expand = expansion(mult = 0, add = 0.01)) +
+    scale_y_discrete(expand = expansion(mult = 0.025, add = 0)) +
     yaxis_title(FALSE) +
     yaxis_text(FALSE) +
     yaxis_ticks(FALSE)
@@ -199,7 +199,7 @@ mcmc_rhat_hist <- function(rhat, ..., binwidth = NULL, bins = NULL, breaks = NUL
     scale_color_diagnostic("rhat") +
     scale_fill_diagnostic("rhat") +
     labs(x = expression(hat(R)), y = NULL) +
-    dont_expand_y_axis(c(0.005, 0)) +
+    dont_expand_y_axis(expansion(mult = 0.005, add = 0)) +
     bayesplot_theme_get() +
     yaxis_title(FALSE) +
     yaxis_text(FALSE) +
@@ -262,7 +262,7 @@ mcmc_neff <- function(ratio, ..., size = NULL) {
       # as.character truncates trailing zeroes, while ggplot default does not
       labels = as.character(breaks),
       limits = c(0, max(1, max_ratio) + 0.05),
-      expand = c(0, 0)) +
+      expand = expansion(0, 0)) +
     bayesplot_theme_get() +
     yaxis_text(FALSE) +
     yaxis_title(FALSE) +
@@ -290,7 +290,7 @@ mcmc_neff_hist <- function(ratio, ..., binwidth = NULL, bins = NULL, breaks = NU
     scale_color_diagnostic("neff_ratio") +
     scale_fill_diagnostic("neff_ratio") +
     labs(x = expression(N[eff]/N), y = NULL) +
-    dont_expand_y_axis(c(0.005, 0)) +
+    dont_expand_y_axis(expansion(mult = 0.005, add = 0)) +
     yaxis_title(FALSE) +
     yaxis_text(FALSE) +
     yaxis_ticks(FALSE) +
@@ -541,9 +541,7 @@ drop_NAs_and_warn <- function(x) {
       bayesplot_theme_get()
     if (style == "bar") {
       graph <- graph +
-        geom_bar(
-          position = "identity",
-          stat = "identity",
+        geom_col(
           linewidth = 0.2,
           fill = get_color("l"),
           color = get_color("lh"),
@@ -574,7 +572,7 @@ drop_NAs_and_warn <- function(x) {
       scale_x_continuous(
         limits = c(-0.5, lags + 0.5),
         breaks = function(x) as.integer(pretty(x, n = 3)),
-        expand = c(0, 0)
+        expand = expansion(0, 0)
       ) +
       labs(x = "Lag", y = "Autocorrelation") +
       force_axes_in_facets()